requirements.txt
selenium==4.11.2
webdriver-manager==3.5.3
mysql-connector-python